+++ migrations/gophermart/000003_create_balance_transactions_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE gophermart.balance_transactions (
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL REFERENCES gophermart.users(id) ON DELETE CASCADE,
     order_number VARCHAR(255) NOT NULL,
-    sum DECIMAL(10, 2) NOT NULL,
+    sum BIGINT NOT NULL,
     processed_at TIMESTAMP WITH TIME ZONE DEFAULT NOW() NOT NULL
 );
